moved several methods from the utils folder to their more feature specific folder
