+++ src/applications/vaos/appointment-list/components/cards/confirmed/ConfirmedAppointmentListItem.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import classNames from 'classnames';
-import moment from '../../../../utils/moment-tz';
-import { formatFacilityAddress } from '../../../../utils/formatters';
+import moment from '../../../../lib/moment-tz';
+import { formatFacilityAddress } from '../../../../services/location/transformers';
 import {
   APPOINTMENT_STATUS,
   PURPOSE_TEXT,
