+++ src/applications/vaos/appointment-list/redux/reducer.js
@@ -32,10 +32,19 @@ import {
   FETCH_STATUS,
   APPOINTMENT_STATUS,
   EXPRESS_CARE,
-  WEEKDAY_INDEXES,
 } from '../../utils/constants';
 import { distanceBetween } from '../../utils/address';
 import { getFacilityIdFromLocation } from '../../services/location';
+
+const WEEKDAY_INDEXES = {
+  SUNDAY: 0,
+  MONDAY: 1,
+  TUESDAY: 2,
+  WEDNESDAY: 3,
+  THURSDAY: 4,
+  FRIDAY: 5,
+  SATURDAY: 6,
+};
 
 const initialState = {
   pending: null,
