+++ src/applications/vaos/new-appointment/components/ConfirmationPage/ConfirmationDirectScheduleInfo.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import moment from '../../../utils/moment-tz.js';
+import moment from '../../../lib/moment-tz.js';
 import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 import FacilityAddress from '../../../components/FacilityAddress';
 import AddToCalendar from '../../../components/AddToCalendar';
-import { formatFacilityAddress } from '../../../utils/formatters';
+import { formatFacilityAddress } from '../../../services/location/transformers';
 import {
   getTimezoneAbbrBySystemId,
   getTimezoneBySystemId,
