+++ src/applications/vaos/tests/appointment-list/components/PastAppointmentsList/PastAppointmentsDateDropdown.unit.spec.jsx
@@ -4,7 +4,7 @@ import sinon from 'sinon';
 import { mount } from 'enzyme';
 import moment from 'moment';
 import PastAppointmentsDateDropdown from '../../../../appointment-list/components/PastAppointmentsList/PastAppointmentsDateDropdown';
-import { getPastAppointmentDateRangeOptions } from '../../../../utils/appointment';
+import { getPastAppointmentDateRangeOptions } from '../../../../appointment-list/components/PastAppointmentsList';
 
 const dateRanges = getPastAppointmentDateRangeOptions(moment('2020-02-02'));
