+++ src/applications/vaos/tests/appointment-list/components/PastAppointmentsList/index.unit.spec.jsx
@@ -15,7 +15,9 @@ import {
 } from '../../../mocks/helpers';
 import { renderWithStoreAndRouter } from '../../../mocks/setup';
 
-import PastAppointmentsList from '../../../../appointment-list/components/PastAppointmentsList';
+import PastAppointmentsList, {
+  getPastAppointmentDateRangeOptions,
+} from '../../../../appointment-list/components/PastAppointmentsList';
 
 const initialState = {
   featureToggles: {
@@ -365,5 +367,41 @@ describe('VAOS integration: past appointments', () => {
       secondDate.format('dddd, MMMM D, YYYY [at] h:mm a'),
       thirdDate.format('dddd, MMMM D, YYYY [at] h:mm a [UTC UTC]'),
     ]);
+  });
+  describe('getPastAppointmentDateRangeOptions', () => {
+    const ranges = getPastAppointmentDateRangeOptions(moment('2020-02-02'));
+    it('should return 6 correct date ranges for dropdown', () => {
+      expect(ranges.length).to.equal(6);
+
+      expect(ranges[0].value).to.equal(0);
+      expect(ranges[0].label).to.equal('Past 3 months');
+      expect(ranges[0].startDate).to.include('2019-11-02T00:00:00');
+      expect(ranges[0].endDate).to.include('2020-02-02T00:00:00');
+
+      expect(ranges[1].value).to.equal(1);
+      expect(ranges[1].label).to.equal('Sept. 2019 – Nov. 2019');
+      expect(ranges[1].startDate).to.include('2019-09-01T00:00:00');
+      expect(ranges[1].endDate).to.include('2019-11-30T23:59:59');
+
+      expect(ranges[2].value).to.equal(2);
+      expect(ranges[2].label).to.equal('June 2019 – Aug. 2019');
+      expect(ranges[2].startDate).to.include('2019-06-01T00:00:00');
+      expect(ranges[2].endDate).to.include('2019-08-31T23:59:59');
+
+      expect(ranges[3].value).to.equal(3);
+      expect(ranges[3].label).to.equal('March 2019 – May 2019');
+      expect(ranges[3].startDate).to.include('2019-03-01T00:00:00');
+      expect(ranges[3].endDate).to.include('2019-05-31T23:59:59');
+
+      expect(ranges[4].value).to.equal(4);
+      expect(ranges[4].label).to.equal('All of 2020');
+      expect(ranges[4].startDate).to.include('2020-01-01T00:00:00');
+      expect(ranges[4].endDate).to.include('2020-02-02T00:00:00');
+
+      expect(ranges[5].value).to.equal(5);
+      expect(ranges[5].label).to.equal('All of 2019');
+      expect(ranges[5].startDate).to.include('2019-01-01T00:00:00');
+      expect(ranges[5].endDate).to.include('2019-12-31T23:59:59');
+    });
   });
 });
