+++ src/applications/vaos/tests/components/AddToCalendar.unit.spec.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import { shallow } from 'enzyme';
-import moment from '../../utils/moment-tz.js';
+import moment from '../../lib/moment-tz.js';
 
-import AddToCalendar from '../../components/AddToCalendar';
+import AddToCalendar, { generateICS } from '../../components/AddToCalendar';
 
 describe('VAOS <AddToCalendar>', () => {
   describe('Add VA appointment to calendar', () => {
@@ -89,6 +89,75 @@ describe('VAOS <AddToCalendar>', () => {
     Object.defineProperty(window.navigator, 'msSaveOrOpenBlob', {
       value: oldValue,
       writable: true,
+    });
+  });
+  describe('generateICS', () => {
+    const now = moment();
+    it('should generate valid ICS calendar commands', () => {
+      const momentDate = moment(now);
+      const dtStamp = momentDate.format('YYYYMMDDTHHmmss');
+      const dtStart = momentDate.format('YYYYMMDDTHHmmss');
+      const dtEnd = momentDate
+        .clone()
+        .add(60, 'minutes')
+        .format('YYYYMMDDTHHmmss');
+
+      const ics = generateICS(
+        'Community Care',
+        '. ',
+        'Address 1 City, State Zip',
+        dtStart,
+        dtEnd,
+      );
+      expect(ics).to.contain('BEGIN:VCALENDAR');
+      expect(ics).to.contain('VERSION:2.0');
+      expect(ics).to.contain('PRODID:VA');
+      expect(ics).to.contain('BEGIN:VEVENT');
+      expect(ics).to.contain('UID:');
+      expect(ics).to.contain('SUMMARY:Community Care');
+      expect(ics).to.contain('DESCRIPTION:. ');
+      expect(ics).to.contain('LOCATION:Address 1 City, State Zip');
+      expect(ics).to.contain(`DTSTAMP:${dtStamp}`);
+      expect(ics).to.contain(`DTSTART:${dtStart}`);
+      expect(ics).to.contain(`DTEND:${dtEnd}`);
+      expect(ics).to.contain('END:VEVENT');
+      expect(ics).to.contain('END:VCALENDAR');
+    });
+    it('should properly chunk long descriptions', () => {
+      const momentDate = moment(now);
+      const dtStamp = momentDate.format('YYYYMMDDTHHmmss');
+      const dtStart = momentDate.format('YYYYMMDDTHHmmss');
+      const dtEnd = momentDate
+        .clone()
+        .add(60, 'minutes')
+        .format('YYYYMMDDTHHmmss');
+      const description = `Testing long line descriptions
+Testing long descriptions Testing long descriptions Testing long descriptions
+Testing long descriptions Testing long descriptions Testing long descriptions
+Testing long descriptions`;
+
+      const ics = generateICS(
+        'Community Care',
+        description,
+        'Address 1 City, State Zip',
+        dtStart,
+        dtEnd,
+      );
+      expect(ics).to.contain('BEGIN:VCALENDAR');
+      expect(ics).to.contain('VERSION:2.0');
+      expect(ics).to.contain('PRODID:VA');
+      expect(ics).to.contain('BEGIN:VEVENT');
+      expect(ics).to.contain('UID:');
+      expect(ics).to.contain('SUMMARY:Community Care');
+      expect(ics).to.contain(
+        'DESCRIPTION:Testing long line descriptions\\nTesting long descriptions Test\r\n\ting long descriptions Testing long descriptions\\nTesting long descriptions\r\n\t Testing long descriptions Testing long descriptions\\nTesting long descrip\r\n\ttions',
+      );
+      expect(ics).to.contain('LOCATION:Address 1 City, State Zip');
+      expect(ics).to.contain(`DTSTAMP:${dtStamp}`);
+      expect(ics).to.contain(`DTSTART:${dtStart}`);
+      expect(ics).to.contain(`DTEND:${dtEnd}`);
+      expect(ics).to.contain('END:VEVENT');
+      expect(ics).to.contain('END:VCALENDAR');
     });
   });
 });
