+++ src/applications/vaos/tests/new-appointment/components/ClinicChoicePage.unit.spec.jsx
@@ -4,7 +4,10 @@ import sinon from 'sinon';
 import { mount } from 'enzyme';
 
 import { selectRadio } from 'platform/testing/unit/schemaform-utils.jsx';
-import { ClinicChoicePage } from '../../../new-appointment/components/ClinicChoicePage';
+import {
+  ClinicChoicePage,
+  formatTypeOfCare,
+} from '../../../new-appointment/components/ClinicChoicePage';
 import { FETCH_STATUS } from '../../../utils/constants';
 
 describe('VAOS <ClinicChoicePage>', () => {
@@ -188,5 +191,18 @@ describe('VAOS <ClinicChoicePage>', () => {
     expect(form.find('.usa-input-error').length).to.equal(0);
     expect(routeToNextAppointmentPage.called).to.be.true;
     form.unmount();
+  });
+
+  describe('formatTypeOfCare', () => {
+    it('should not lower case MOVE', () => {
+      const result = formatTypeOfCare('MOVE! weight management');
+
+      expect(result).to.equal('MOVE! weight management');
+    });
+    it('should lower case regular types of care', () => {
+      const result = formatTypeOfCare('Primary care');
+
+      expect(result).to.equal('primary care');
+    });
   });
 });
