+++ src/applications/vaos/tests/utils/formatters.unit.spec.js
@@ -1,26 +1,8 @@
 import { expect } from 'chai';
 
-import {
-  formatTypeOfCare,
-  titleCase,
-  sentenceCase,
-  lowerCase,
-} from '../../utils/formatters';
+import { titleCase, sentenceCase, lowerCase } from '../../utils/formatters';
 
 describe('VAOS formatters', () => {
-  describe('formatTypeOfCare', () => {
-    it('should not lower case MOVE', () => {
-      const result = formatTypeOfCare('MOVE! weight management');
-
-      expect(result).to.equal('MOVE! weight management');
-    });
-    it('should lower case regular types of care', () => {
-      const result = formatTypeOfCare('Primary care');
-
-      expect(result).to.equal('primary care');
-    });
-  });
-
   describe('titleCase', () => {
     it('should return capitalize the 1st letter of each word in a sentence', () => {
       expect(titleCase('THE cOw jumpeD over the moon')).to.equal(
