+++ src/applications/vaos/utils/scrollAndFocus.js
@@ -1,17 +1,8 @@
-import { useEffect } from 'react';
+// todo: potential refactor opportunity
+
 import { focusElement } from 'platform/utilities/ui';
 
 export function scrollAndFocus(selector = 'h1,.loading-indicator') {
   const el = document.querySelector(selector);
   focusElement(el);
-}
-
-export function useScrollAndFocus() {
-  useEffect(() => {
-    if (window.History) {
-      window.History.scrollRestoration = 'manual';
-    }
-
-    scrollAndFocus();
-  });
 }
